+++ packages/core/src/components/sections/RecommendationShelf/index.ts
@@ -1,1 +1,7 @@
-export * from './RecommendationShelf'
+export { RecommendationShelf } from './RecommendationShelf'
+export type {
+  ItemContext,
+  RecommendationProductCardMapper,
+  RecommendationShelfProps,
+  RecommendationType,
+} from './RecommendationShelf.types'